requirements.txt
requests[security]
celery
mediacloud
raven
blinker
mediameter-cliff
